Adding separate injection point for v4Signer
