+++ index.d.ts
@@ -56,7 +56,6 @@ import {
   ResurrectEvent,
   BasicAuth,
 } from './lib/pool';
-import AwsV4Signer from './lib/aws/AwsV4Signer';
 import Serializer from './lib/Serializer';
 import Helpers from './lib/Helpers';
 import * as errors from './lib/errors';
@@ -5379,5 +5378,4 @@ export {
   ClientOptions,
   NodeOptions,
   ClientExtendsCallbackOptions,
-  AwsV4Signer,
 };
