+++ index.js
@@ -47,7 +47,6 @@ if (clientVersion.includes('-')) {
   // clean prerelease
   clientVersion = clientVersion.slice(0, clientVersion.indexOf('-')) + 'p';
 }
-const AwsV4Signer = require('./lib/aws/AwsV4Signer');
 
 const kInitialOptions = Symbol('opensearchjs-initial-options');
 const kChild = Symbol('opensearchjs-child');
@@ -349,5 +348,4 @@ module.exports = {
   Serializer,
   events,
   errors,
-  AwsV4Signer,
 };
