+++ index.mjs
@@ -38,4 +38,3 @@ export const Connection = mod.Connection;
 export const Serializer = mod.Serializer;
 export const events = mod.events;
 export const errors = mod.errors;
-export const awsV4Signer = mod.AwsV4Signer;
